backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
requests==2.31.0
beautifulsoup4==4.12.2
PyJWT==2.8.0
python-dotenv==1.0.0
selenium==4.18.1
playwright==1.42.0
spacy==3.7.4
langchain==0.1.12
openai==1.12.0
